+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ---
 
+#### 17.07.2018
+
+* [Rust] Code of Kutulu - Closest explorer with a HashMap (CodinGame)
+
 #### 16.07.2018
 
 * [Rust] Code of Kutulu - Units in a HashMap (CodinGame)